refactor(checkbox): extract checkbox to its own component
add a showCheckbox prop to the Input component, so it can be used with or without the checkbox
