+++ src/app/components/Checkbox/checkbox.module.scss
@@ -1,0 +1,44 @@
+@use '../../../assets/styles/abstracts' as *;
+
+.checkbox-container {
+  position: relative;
+}
+
+.checkbox-mark {
+  width: 1.5rem;
+  height: 1.5rem;
+  border-radius: 50%;
+  opacity: 0;
+  transition: 200ms opacity linear;
+
+  &.active {
+    opacity: 1;
+    background: linear-gradient(145deg, $gradient-1 0%, $gradient-2 100%);
+  }
+}
+
+.checkbox {
+  appearance: none;
+  margin: 0;
+  width: 1.5rem;
+  height: 1.5rem;
+  border: 1px solid #e3e4f1;
+  border-radius: 50%;
+  position: absolute;
+  top: 0;
+  left: 0;
+  display: grid;
+  place-content: center;
+
+  &:checked {
+    border: none;
+  }
+
+  &::before {
+    content: '';
+  }
+
+  &:checked::before {
+    content: url('../../../assets/images/icon-check.svg');
+  }
+}
