+++ src/app/components/Input/input.module.scss
@@ -1,0 +1,15 @@
+@use '../../../assets/styles/abstracts' as *;
+
+.input-container {
+  background-color: #fff;
+  padding: 0.5rem 1rem;
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  border-radius: 0.3125rem;
+
+  .textinput {
+    border: none;
+    outline: none;
+  }
+}
